# examples/user-service/gen/openapi/v1.swagger.json
{
  "consumes": [
    "application/json"
  ],
  "definitions": {
    "protobufAny": {
      "additionalProperties": {},
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "type": "object"
    },
    "rpcStatus": {
      "properties": {
        "code": {
          "format": "int32",
          "type": "integer"
        },
        "details": {
          "items": {
            "$ref": "#/definitions/protobufAny",
            "type": "object"
          },
          "type": "array"
        },
        "message": {
          "type": "string"
        }
      },
      "type": "object"
    },
...
    "v1GetUserResponse": {
      "properties": {
        "user": {
          "$ref": "#/definitions/v1User"
        }
      },
      "type": "object"
    },
...
    "v1User": {
      "properties": {
        "email": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      },
      "type": "object"
    },
...
  },
  "info": {
    "title": "api/v1/user/service.proto (v1)",
    "version": "version not set"
  },
  "paths": {
...
    "/v1/users/{id}": {
      "get": {
        "operationId": "UserService_GetUser",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "UserService"
        ]
      }
    }
  },
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
...
}